# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "paths": {
...
        "/api/dashboard": {
            "post": {
                "tags": [
                    "Authentication"
                ],
                "summary": "Dashboard",
                "description": "Customer Orders",
                "operationId": "dashboard",
                "requestBody": {
                    "description": "Pet object that needs to be added to the store",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "customer_email"
                                ],
                                "properties": {
                                    "customer_email": {
                                        "description": "Customer Email",
                                        "type": "string",
                                        "example": "gomathi@gmail.com"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "400": {
                        "description": "Invalid input"
                    }
                }
            }
        },
...
    },
...
    "security": {
        "passport": {
            "type": "oauth2",
            "description": "Laravel passport oauth2 security.",
            "in": "header",
            "scheme": "https",
            "flows": {
                "password": {
                    "authorizationUrl": "https://npcoder.com/oauth/authorize",
                    "tokenUrl": "https://npcoder.com/oauth/token",
                    "refreshUrl": "https://npcoder.com/token/refresh",
                    "scopes": []
                }
            }
        }
    }
}